removing healthit product mock
