+++ endpointmanager/pkg/chplquerier/productquerier_test.go
@@ -13,10 +13,7 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/endpointmanager"
-	"github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/endpointmanager/mock"
 	th "github.com/onc-healthit/lantern-back-end/endpointmanager/pkg/testhelper"
-
-	logtest "github.com/sirupsen/logrus/hooks/test"
 
 	"github.com/spf13/viper"
 )
@@ -292,143 +289,6 @@ func Test_prodNeedsUpdate(t *testing.T) {
 	th.Assert(t, origErr.Error() == expectedErrorStr, fmt.Sprintf("For 'prodNeedsUpdate' using %s, expected error\n%v\nAnd got error\n%v", name, expectedErrorStr, origErr))
 }
 
-func Test_persistProduct(t *testing.T) {
-	var err error
-	store := mock.NewBasicMockHealthITProductStore()
-
-	var ctx context.Context
-	var cancel context.CancelFunc
-
-	prod := testCHPLProd
-	hitp := testHITP
-
-	// check that ended context when no element in store fails as expected
-	ctx, cancel = context.WithCancel(context.Background())
-	cancel()
-	err = persistProduct(ctx, store, &prod)
-	th.Assert(t, len(store.(*mock.BasicMockStore).HealthITProductData) == 0, "should not have stored data")
-	th.Assert(t, errors.Cause(err) == context.Canceled, "should have errored out with root cause that the context was canceled")
-
-	// reset context
-	ctx = context.Background()
-
-	// check that new item is stored
-	err = persistProduct(ctx, store, &prod)
-	th.Assert(t, err == nil, err)
-	th.Assert(t, len(store.(*mock.BasicMockStore).HealthITProductData) == 1, "did not store data as expected")
-	th.Assert(t, hitp.Equal(store.(*mock.BasicMockStore).HealthITProductData[0]), "stored data does not equal expected store data")
-
-	// check that newer updated item replaces item
-	prod.Edition = "2015"
-	hitp.CertificationEdition = "2015"
-	err = persistProduct(ctx, store, &prod)
-	th.Assert(t, err == nil, err)
-	th.Assert(t, len(store.(*mock.BasicMockStore).HealthITProductData) == 1, "did not store data as expected")
-	th.Assert(t, hitp.Equal(store.(*mock.BasicMockStore).HealthITProductData[0]), "stored data does not equal expected store data")
-
-	// check that older updated item does not replace item
-	prod.Edition = "2014"
-	hitp.CertificationEdition = "2015" // keeping 2015
-	err = persistProduct(ctx, store, &prod)
-	th.Assert(t, err == nil, err)
-	th.Assert(t, len(store.(*mock.BasicMockStore).HealthITProductData) == 1, "did not store data as expected")
-	th.Assert(t, hitp.Equal(store.(*mock.BasicMockStore).HealthITProductData[0]), "stored data does not equal expected store data")
-
-	// check that malformed product throws error
-	prod.APIDocumentation = "170.315 (g)(7),http://carefluence.com/Carefluence-OpenAPI-Documentation.html"
-	err = persistProduct(ctx, store, &prod)
-	th.Assert(t, err != nil, "expected error parsing product")
-
-	// check that ambiguous update throws error
-	prod = testCHPLProd
-	prod.Edition = "2015" // same date as what is in store
-	prod.CertificationStatus = "Retired"
-	err = persistProduct(ctx, store, &prod)
-	th.Assert(t, err != nil, "expected error updating product")
-
-	// check that error adding to store throws error
-	prod = testCHPLProd
-	prod.Product = "A new product name"
-	addFn := store.(*mock.BasicMockStore).AddHealthITProductFn
-	store.(*mock.BasicMockStore).AddHealthITProductFn = func(_ context.Context, _ *endpointmanager.HealthITProduct) error {
-		return errors.New("add healthitproduct test error")
-	}
-	err = persistProduct(ctx, store, &prod)
-	th.Assert(t, errors.Cause(err).Error() == "add healthitproduct test error", "expected error adding product")
-	store.(*mock.BasicMockStore).AddHealthITProductFn = addFn
-
-	// check that error updating to store throws error
-	prod = testCHPLProd
-	prod.Edition = "2016"
-	updateFn := store.(*mock.BasicMockStore).UpdateHealthITProductFn
-	store.(*mock.BasicMockStore).UpdateHealthITProductFn = func(_ context.Context, _ *endpointmanager.HealthITProduct) error {
-		return errors.New("update healthitproduct test error")
-	}
-	err = persistProduct(ctx, store, &prod)
-	th.Assert(t, errors.Cause(err).Error() == "update healthitproduct test error", "expected error updating product")
-	store.(*mock.BasicMockStore).UpdateHealthITProductFn = updateFn
-
-}
-
-func Test_persistProducts(t *testing.T) {
-	var err error
-	store := mock.NewBasicMockHealthITProductStore()
-
-	ctx := context.Background()
-
-	// standard persist
-
-	prod1 := testCHPLProd
-	prod2 := testCHPLProd
-	prod2.Product = "another prod"
-
-	prodList := chplCertifiedProductList{Results: []chplCertifiedProduct{prod1, prod2}}
-
-	err = persistProducts(ctx, store, &prodList)
-	th.Assert(t, err == nil, err)
-
-	th.Assert(t, len(store.(*mock.BasicMockStore).HealthITProductData) == 2, "did not persist two products as expected")
-	th.Assert(t, store.(*mock.BasicMockStore).HealthITProductData[0].Name == testCHPLProd.Product, "Did not store first product as expected")
-	th.Assert(t, store.(*mock.BasicMockStore).HealthITProductData[1].Name == "another prod", "Did not store second product as expected")
-
-	// persist with errors
-
-	hook := logtest.NewGlobal()
-	store = mock.NewBasicMockHealthITProductStore()
-
-	prod2.APIDocumentation = "170.315 (g)(7),http://carefluence.com/Carefluence-OpenAPI-Documentation.html"
-	expectedErr := "retreiving the API URL from the health IT product API documentation list failed: unexpected format for api doc string"
-	prodList = chplCertifiedProductList{Results: []chplCertifiedProduct{prod1, prod2}}
-
-	err = persistProducts(ctx, store, &prodList)
-	// don't expect the function to return with errors
-	th.Assert(t, err == nil, err)
-	// only expect one item to be stored
-	th.Assert(t, len(store.(*mock.BasicMockStore).HealthITProductData) == 1, "did not persist one product as expected")
-	th.Assert(t, store.(*mock.BasicMockStore).HealthITProductData[0].Name == testCHPLProd.Product, "Did not store first product as expected")
-	// expect presence of a log message
-	found := false
-	for i := range hook.Entries {
-		if hook.Entries[i].Message == expectedErr {
-			found = true
-			break
-		}
-	}
-	th.Assert(t, found, "expected an error to be logged")
-
-	// persist when context has ended
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	store = mock.NewBasicMockHealthITProductStore()
-
-	prod2 = testCHPLProd
-	prod2.Product = "another prod"
-
-	err = persistProducts(ctx, store, &prodList)
-	th.Assert(t, errors.Cause(err) == context.Canceled, "expected persistProducts to error out due to context ending")
-}
-
 func Test_getProductJSON(t *testing.T) {
 	var err error
 	var tc *th.TestClient
@@ -501,73 +361,6 @@ func Test_getProductJSON(t *testing.T) {
 	default:
 		t.Fatal("Expected url error")
 	}
-}
-
-func Test_GetCHPLProducts(t *testing.T) {
-	var err error
-	var tc *th.TestClient
-	var ctx context.Context
-	var store endpointmanager.HealthITProductStore
-
-	// basic test
-
-	// mock JSON includes 201 product entries, but w duplicates, the number stored is 168.
-	expectedProdsStored := 168
-
-	tc, err = basicTestClient()
-	th.Assert(t, err == nil, err)
-	defer tc.Close()
-
-	ctx = context.Background()
-
-	store = mock.NewBasicMockHealthITProductStore()
-
-	err = GetCHPLProducts(ctx, store, &(tc.Client))
-	th.Assert(t, err == nil, err)
-	actualProdsStored := len(store.(*mock.BasicMockStore).HealthITProductData)
-	th.Assert(t, actualProdsStored == expectedProdsStored, fmt.Sprintf("Expected %d products stored. Actually had %d products stored.", expectedProdsStored, actualProdsStored))
-
-	// test context ended
-	// also checks what happens when an http request fails
-
-	tc, err = basicTestClient()
-	th.Assert(t, err == nil, err)
-	defer tc.Close()
-
-	ctx, cancel := context.WithCancel(context.Background())
-	cancel()
-
-	store = mock.NewBasicMockHealthITProductStore()
-
-	err = GetCHPLProducts(ctx, store, &(tc.Client))
-	switch reqErr := errors.Cause(err).(type) {
-	case *url.Error:
-		th.Assert(t, reqErr.Err == context.Canceled, "Expected error stating that context was canceled")
-	default:
-		t.Fatal("Expected context canceled error")
-	}
-
-	// test with malformed json
-
-	malformedJSON := `
-		"asdf": [
-		{}]}
-		`
-	tc = th.NewTestClientWithResponse([]byte(malformedJSON))
-	defer tc.Close()
-
-	ctx = context.Background()
-
-	store = mock.NewBasicMockHealthITProductStore()
-
-	err = GetCHPLProducts(ctx, store, &(tc.Client))
-	switch errors.Cause(err).(type) {
-	case *json.SyntaxError:
-		// ok
-	default:
-		t.Fatal("Expected JSON syntax error")
-	}
-
 }
 
 func basicTestClient() (*th.TestClient, error) {
